pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "File difference calculator"
authors = [
    { name = "Alexandra Minina", email = "i@minina-work.ru" },
    { name = "Hexlet", email = "hexlet@hexlet.io" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=9.0.2",
    "pytest-cov>=4.0.0",
    "ruff>=0.15.8",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
test = ["pytest"]

[project.scripts]
gendiff = "gendiff.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]
